Tried adding rb.AddForce
Tried a whole new approach with AddForce instead of setting the velocity manually. Should in theory be easier to work with (especially in regards to torque) but it's REALLY finicky and prone to a positive feedback loop on the velocity, making the plane shoot out of the sky with no warning. This is likely due to angle-of-attack calculations doing something wrong.